+++ DotDoc.EntityFrameworkCore.Extensions.Tests/ExecuteUpdateExtensions/ExecuteUpdateGetRowsTests.cs
@@ -32,7 +32,7 @@ public class ExecuteUpdateGetRowsTests
         string paramName = "optionsBuilder";
 
         // ACT / ASSERT
-        ArgumentNullException e = Assert.ThrowsException<ArgumentNullException>(() => optionsBuilder!.UseExecuteUpdateExtensions(), "Missing exception");
+        ArgumentNullException e = Assert.ThrowsExactly<ArgumentNullException>(() => _ = optionsBuilder!.UseExecuteUpdateExtensions(), "Missing exception");
         Assert.AreEqual(paramName, e.ParamName, "Invalid parameter name");
     }