+++ DotDoc.EntityFrameworkCore.Extensions.Tests/UniqueConstraintExtensions/UniqueConstraintExceptionTests.cs
@@ -28,7 +28,7 @@ public class UniqueConstraintExceptionTests
         string paramName = "details";
 
         // ACT / ASSERT
-        ArgumentNullException e = Assert.ThrowsException<ArgumentNullException>(() => new UniqueConstraintException(innerException, details!), "Missing exception");
+        ArgumentNullException e = Assert.ThrowsExactly<ArgumentNullException>(() => _ = new UniqueConstraintException(innerException, details!), "Missing exception");
         Assert.AreEqual(paramName, e.ParamName, "Invalid parameter name");
     }